refactor search input to have less inputs
